requirements.txt
discord.py==2.6.4
pymongo==4.15.3
redis[hiredis]==6.4.0